requirements.txt
requests>=2.31.0
websocket-client>=1.6.4
pandas>=2.0.0
scikit-learn>=1.3.0
networkx>=3.1
matplotlib>=3.7.0
seaborn>=0.12.0
numpy>=1.24.0
scipy>=1.10.0
fastapi>=0.104.0
uvicorn>=0.24.0
jinja2>=3.1.0
python-multipart>=0.0.6
websockets>=12.0
folium>=0.14.0
streamlit>=1.28.0
streamlit-folium>=0.13.0
plotly>=5.17.0
python-jose[cryptography]>=3.3.0
authlib>=1.2.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
pydantic>=2.5.0
python-dotenv>=0.21.0
